+++ src/components/calendar/MainCalendar.tsx
@@ -1,10 +1,5 @@
 import * as S from './MainCalendar.style';
-import {
-  Calendar,
-  dayjsLocalizer,
-  Event,
-  ToolbarProps,
-} from 'react-big-calendar';
+import { Calendar, dayjsLocalizer } from 'react-big-calendar';
 import dayjs from 'dayjs';
 import 'dayjs/locale/ko';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
@@ -13,59 +8,13 @@ import {
   IoIosArrowDroprightCircle,
 } from 'react-icons/io';
 import { useMemo } from 'react';
-
-type CustomToolbarProps = Pick<ToolbarProps, 'label' | 'onNavigate'>;
-interface IEventList extends Event {
-  color: IEventColorProps;
-}
-interface IEventColorProps {
-  id: number;
-  bgColor: string;
-  fontColor: string;
-}
+import { CustomToolbarProps, IEventList } from '@/types/calendar';
+import { eventList } from '@/mocks/data/calendar';
 
 dayjs.locale('ko');
 const localizer = dayjsLocalizer(dayjs);
 
 // onSelectSlot: https://jquense.github.io/react-big-calendar/examples/index.html?path=/docs/props--on-select-slot
-
-const arrEventColor: IEventColorProps[] = [
-  {
-    id: 1,
-    bgColor: '#f8f2a2',
-    fontColor: '#000',
-  },
-  {
-    id: 2,
-    bgColor: '#a2a5f8',
-    fontColor: '#fff',
-  },
-  {
-    id: 3,
-    bgColor: '#eaa687',
-    fontColor: '#fff',
-  },
-  {
-    id: 4,
-    bgColor: '#f57897',
-    fontColor: '#fff',
-  },
-  {
-    id: 5,
-    bgColor: '#584b48',
-    fontColor: '#fff',
-  },
-  {
-    id: 6,
-    bgColor: '#dbF0e7',
-    fontColor: '#000',
-  },
-  {
-    id: 7,
-    bgColor: '#5a5a5a',
-    fontColor: '#fff',
-  },
-];
 
 const CustomToolbar = (props: CustomToolbarProps) => {
   const { label, onNavigate } = props;
@@ -104,25 +53,6 @@ const MainCalendar = () => {
     [],
   );
 
-  // 임시 테스트 배열
-  // end: 실제 일정 끝나는 날짜에 + 1 해서 넣어야함. 일정이 지정된 날 전날에 끝남.
-  const testEventsList: IEventList[] = [
-    {
-      title: '일정 테스트',
-      start: new Date(),
-      end: new Date(),
-      color: arrEventColor[0],
-      // allDay: true
-    },
-    {
-      title: '222',
-      start: new Date(),
-      end: new Date(2024, 11, 23),
-      color: arrEventColor[6],
-      // allDay: true
-    },
-  ];
-
   // (event:IEventList, start:Date, end:Date, isSelected:boolean)
   const eventStyleGetter = (event: IEventList) => {
     const style = {
@@ -137,7 +67,7 @@ const MainCalendar = () => {
     <S.Container>
       <Calendar
         localizer={localizer}
-        events={testEventsList}
+        events={eventList}
         startAccessor="start"
         endAccessor="end"
         style={{ height: 500 }}
